requirements.txt
requests
Mastodon.py
beautifulsoup4
